+++ WidgetExtension/PocketCastsWidgetBundle.swift
@@ -6,6 +6,8 @@ struct PocketCastsWidgetBundle: WidgetBundle {
     var body: some Widget {
         NowPlayingWidget()
         UpNextWidget()
+        NowPlayingLockScreenWidget()
+        AppIconWidget()
+        UpNextLockScreenWidget()
     }
 }
-